moved site-specific react components to folder site under react-components
