+++ client/react-components/site/code.jsx
@@ -3,7 +3,7 @@ const PropTypes = require('prop-types');
 const Prism = require('prismjs');
 require('prismjs/components/prism-jsx');
 import 'prismjs/themes/prism.css';
-import '../css/site/mongodb-prism.less';
+import '../../css/site/mongodb-prism.less';
 
 
 class Code extends React.Component {
