+++ client/site/ui-design-system/base-styles/colors/index.jsx
@@ -3,7 +3,7 @@
 // ==================================================
 
 import React from 'react';
-import ColorChip from '../../../../site-components/color-chip';
+import ColorChip from '../../../../react-components/site/color-chip';
 
 const UIColors = () => (
   <div className="wrap">
