+++ client/site/ui-design-system/components/bubbles/index.jsx
@@ -3,7 +3,7 @@
 // ==================================================
 
 import React from 'react';
-import Code from '../../../../site-components/code';
+import Code from '../../../../react-components/site/code';
 
 const UIBubbles = () => (
   <div className="wrap button-ui">
