+++ client/site/ui-design-system/components/buttons/react-buttons.jsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 import { Link } from 'react-router';
-import Code from '../../../../site-components/code';
+import Code from '../../../../react-components/site/code';
 import Button from '../../../../react-components/Button.js';
 import Checkbox from '../../../../react-components/Checkbox.js';
 import { RadioGroup, Radio } from '../../../../react-components/RadioGroup.js';
