+++ client/site/ui-design-system/components/code/react-copyablecommand.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router';
-import Code from '../../../../site-components/code';
+import Code from '../../../../react-components/site/code';
 import CopyableCommand from '../../../../react-components/CopyableCommand.js';
 import Button from '../../../../react-components/Button.js';
 import { RadioGroup, Radio } from '../../../../react-components/RadioGroup.js';
